requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
autopep8==2.0.1
beautifulsoup4==4.11.1
botbuilder==0.0.1
botbuilder-core==4.14.3
botbuilder-schema==4.14.3
botframework-connector==4.14.3
botframework-streaming==4.14.3
bs4==0.0.1
cachetools==5.3.0
cattrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chatbot==1.5.2b0
click==8.1.3
colorama==0.4.5
comtypes==1.1.14
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
decorator==5.1.1
docutils==0.19
exceptiongroup==1.0.0rc9
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-search==1.1.1
googletrans==3.1.0a0
googletranslate==2.1.2
grpcio==1.51.1
gTTS==2.3.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.12.0
hyperframe==5.2.0
idna==2.10
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==1.4.2
keras==2.11.0
keyboard==0.13.5
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kiwisolver==1.4.4
lazy_loader==0.1
libclang==15.0.6.1
librosa==0.10.0
llvmlite==0.39.1
load-dotenv==0.1.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.0
MouseInfo==0.1.3
msal==1.21.0
msgpack==1.0.4
msrest==0.6.21
multidict==6.0.4
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.26.5
opencv-python==4.7.0.68
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
Pillow==9.2.0
platformdirs==3.0.0
playsound==1.3.0
pooch==1.7.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.6.0
PyMsgBox==1.0.9
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyPDF2==2.11.0
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pyttsx3==2.90
pytweening==1.0.4
pywhatkit==5.4
pywikihow==0.5.7
pywin32==304
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-cache==0.9.6
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
selenium==4.1.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.3
SpeechRecognition==3.8.1
srt==3.5.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.12.1
tqdm==4.64.1
transformers==4.26.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
ujson==5.7.0
url-normalize==1.4.3
urllib3==1.26.12
urllib3-secure-extra==0.1.0
vosk==0.3.45
websockets==10.4
Werkzeug==2.2.2
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
